refactor: rename activity timeline
Rename the Activity view to Timeline, as seeing ActivityFragment was upsetting
