+++ app/src/main/java/com/chesire/malime/flow/timeline/TimelineFragment.kt
@@ -1,17 +1,17 @@
-package com.chesire.malime.flow.activity
+package com.chesire.malime.flow.timeline
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.chesire.lifecyklelog.LogLifecykle
-import com.chesire.malime.databinding.FragmentActivityBinding
+import com.chesire.malime.databinding.FragmentTimelineBinding
 import dagger.android.support.DaggerFragment
 
 @LogLifecykle
-class ActivityFragment : DaggerFragment() {
+class TimelineFragment : DaggerFragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = FragmentActivityBinding.inflate(inflater, container, false).root
+    ) = FragmentTimelineBinding.inflate(inflater, container, false).root
 }
