+++ app/src/main/java/com/chesire/malime/injection/androidmodules/FragmentModule.kt
@@ -1,6 +1,6 @@
 package com.chesire.malime.injection.androidmodules
 
-import com.chesire.malime.flow.activity.ActivityFragment
+import com.chesire.malime.flow.timeline.TimelineFragment
 import com.chesire.malime.flow.login.details.DetailsFragment
 import com.chesire.malime.flow.login.syncing.SyncingFragment
 import com.chesire.malime.flow.profile.ProfileFragment
@@ -14,9 +14,6 @@ import dagger.android.ContributesAndroidInjector
 @Suppress("unused")
 @Module
 abstract class FragmentModule {
-    @ContributesAndroidInjector
-    abstract fun contributeActivityFragment(): ActivityFragment
-
     @ContributesAndroidInjector
     abstract fun contributeAnimeFragment(): AnimeFragment
 
@@ -37,4 +34,7 @@ abstract class FragmentModule {
 
     @ContributesAndroidInjector
     abstract fun contributeSyncingFragment(): SyncingFragment
+
+    @ContributesAndroidInjector
+    abstract fun contributeTimelineFragment(): TimelineFragment
 }
