+++ app/src/main/res/values/strings.xml
@@ -4,7 +4,7 @@
     <string name="nav_anime">Anime</string>
     <string name="nav_manga">Manga</string>
     <string name="nav_profile">Profile</string>
-    <string name="nav_activity">Activity</string>
+    <string name="nav_timeline">Timeline</string>
     <string name="nav_settings">Settings</string>
 
     <string name="menu_delete">Delete</string>
